Implement commander combat, navigation, and selection systems
Add the army skill bridge, sweep combat, float health, protocol v6 move planning, commander camera and navigation stabilization, selection UI, diagnostics, tests, data assets, and documentation. Disable client positional move prediction that caused visible pullback.

+++ Source/GuLiStrike/Commander/UI/GuLiCommanderHealthBarRenderer.h
@@ -47,12 +47,12 @@ public:
 	/** Pure production-policy accessors used by focused automation tests. */
 	static bool TestOnly_ShouldDisplayHealthBar(
 		bool bAlive,
-		uint8 Health,
-		uint8 MaxHealth,
+		float Health,
+		float MaxHealth,
 		bool bSelected,
 		float DistanceCentimeters,
 		float MaximumDistanceCentimeters);
-	static float TestOnly_CalculateHealthFraction(uint8 Health, uint8 MaxHealth);
+	static float TestOnly_CalculateHealthFraction(float Health, float MaxHealth);
 	static FVector2D TestOnly_CalculateWorldSizeCentimeters(
 		float DistanceCentimeters,
 		float HorizontalFieldOfViewDegrees,
@@ -77,12 +77,12 @@ private:
 
 	static bool ShouldDisplayHealthBar(
 		bool bAlive,
-		uint8 Health,
-		uint8 MaxHealth,
+		float Health,
+		float MaxHealth,
 		bool bSelected,
 		float DistanceCentimeters,
 		float MaximumDistanceCentimeters);
-	static float CalculateHealthFraction(uint8 Health, uint8 MaxHealth);
+	static float CalculateHealthFraction(float Health, float MaxHealth);
 	static FVector2D CalculateWorldSizeCentimeters(
 		float DistanceCentimeters,
 		float HorizontalFieldOfViewDegrees,
