+++ Source/GuLiStrike/Commander/UI/GuLiCommanderUnitTypeSummary.h
@@ -10,8 +10,8 @@ struct GULISTRIKE_API FGuLiCommanderUnitTypeSummary
 {
 	int32 SelectedCount = 0;
 	int32 AliveCount = 0;
-	int32 TotalHealth = 0;
-	int32 TotalMaxHealth = 0;
+	float TotalHealth = 0.0f;
+	float TotalMaxHealth = 0.0f;
 	bool bSyncing = false;
 	FText CommandStatus;
